YH-698: changed the title text and the color of unsubscribe button
