+++ public/locales/ru/subscription.json
@@ -6,7 +6,7 @@
   "tariff": {
     "free": "Бесплатный",
     "base": "Базовый",
-    "premium": "Премиум"
+    "premium": "Участник сообщества"
   },
   "daysLeft_one": "Осталось {{count}} день",
   "daysLeft_few": "Осталось {{count}} дня",
@@ -22,7 +22,7 @@
     "second": {
       "question": "В чем преимущества членства?",
       "answer": "Членство открывает полный доступ ко всем материалам, эксклюзивным подборкам и активностям, а также даёт возможность поддерживать развитие сообщества и платформы."
-    }, 
+    },
     "third": {
       "question": "Как отменить подписку?",
       "answer": "Вы можете отменить подписку в любое время через ваш аккаунт на платформе. Доступ к платным материалам будет прекращен с конца текущего месяца."
