+++ src/widgets/Profile/ui/SubscriptionTab/ui/PremiumSubscriptionTab/PremiumSubscriptionTab.module.css
@@ -5,7 +5,7 @@
 .svg {
   width: 32px;
   height: 33px;
-  fill: var( --color-green-750);
+  fill: var(--color-yellow-900);
 }
 
 .card-title {
